// front/src/app/model/enums.ts
// Products, must be the same strings as in sevdesk

export enum ProductsEnum {
    W2B = '2-Kg-Beutel-Würfeleis',
    W5B = '5-Kg-Beutel-Würfeleis',
    W15B = '15-Kg-Beutel-Würfeleis',
    WL = 'Würfeleis',
    C2B = "2-Kg-Beutel-Crushed-Eis",
    C5B = '5-Kg-Beutel-Crushed-Eis',
    C15B = '15-Kg-Beutel-Crushed-Eis',
    CL = 'Crushed-Eis'

};